Revert "スピードの調整は、これから"
This reverts commit 06c3e82612f462a824d2f17df52d90b62a02095a.

+++ EatingBuild/Assets/Script/NPCMove.cs
@@ -11,7 +11,7 @@ public class NPCMove : MonoBehaviour
 {
 
     [SerializeField] private NPCEatObjectScript npceat;
-    private float p;
+    private int p;
     [SerializeField] GameManager gameManager;
 
     //位置の基準になるオブジェクトのTransformを収める
